Refactoring: enrichment.py from query_engine.py
